==> 1_Introduction - Business_Understanding.ipynb <==
# UVA.Capstone1_webscrapper
### CS 5010, Summer 2020

* Alfred Hull, adh8mb
* Farhan Kanani, fk3ak
* Dara Maguire, dkm2bd
* Francisco Estrada, fge8tj
* Niraja Bohidar, nb4dg

### Phase I. Business Understanding.

### Background

Severe Acute Respiratory Syndrome Coronavirus 2 (SARS-CoV-2), also known as COVID-19, has manifested into a pandemic that has spread to 188 countries and territories with more than 17.5 million cases as of August 2020. In the United States alone there have 4.68 million confirmed cases and rising.

Though the complete means of the method of transmission is still unclear, several scientists postulate that transmission of COVID-19 occurs primarily in droplets through the air. In order to combat this, health officials have recommended that people wear masks and face shields, and maintain social distancing of at least 6 feet. However, many researchers believe that in order to truly help prevent transmission, we need to protect the air. 

One aspect of air quality that seems to have been overlooked is humidity. Infection control specialist Stephanie Taylor identified that midrange humidity levels have been able to help prevent the transmission of airborne pathogens. When air is too dry, large air droplets dry out to become smaller droplets. In moister air, airborne viruses like COVID-19, don't seem to be as virulent, though the mechanism is not fully understood. Though creating a more humid environment isn't a complete solution to prevent transmission, it seems to be another step in creating a safer environment. 
 
Healthcare professionals, businesses, household owners, and the general public. It would be in their best interests to consider implementing these precautions in order to deter the spread of COVID-19.

### Who might care:

Healthcare professionals, businesses, household owners, and the general public. It would be in their best interests to consider implementing these precautions in order to deter the spread of COVID-19.

### Set-Objective:

The objective is to corroborate the evidence that humidity may have a role in reducing transmission by looking at the humidity in each state and the number of cases in those states. 

Should our data corroborate these findings, it will be further justification for businesses to maintain a midrange (40%-60%) humitity level in their establishments. 
 
### Cost and benefits:

Though businesses may invest more in quality humidifiers to maintain this level of humidity, it may be a necessity in order create a safer environment for people.
